+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Javis.jl - Changelog
 
+## v0.6.1 (7th of August 2021)
+- Add shorthands for basic shapes
+  - New functions `JBox, JCircle, JEllipse, JLine, JPoly, JRect, JStar, @JShape` 
+- added support for `rescale_factor` keyword in `render` function
+- Docstring improvements to `translate`
+
+
 ## v0.6.0 (3rd of August 2021)
 - Added layers see `@JLayer`
 
@@ -11,7 +18,7 @@
 - Morphing mutates the object function 
 
 ## v0.5.2
-- Added support for local network live streaming
+- added support for local network live streaming
 
 ## v0.5.1
 - added support for Pluto notebooks